// src/enums.ts
export enum CellColor {
  WHITE = 'white',
  BLACK = 'black',
}

export enum FigureTeam {
  WHITE = 'white',
  BLACK = 'black',
}

export enum HighlightType {
  NONE = 'none',
  SELECTED = 'selected',
  STEP = 'step',
  KILL = 'kill',
}

export enum FigureType {
  PAWN = 'pawn',
  KING = 'king',
  QUEEN = 'queen',
  KNIGHT = 'knight',
  BISHOP = 'bishop',
  ROOK = 'rook',
}